Moved from the multiple activity to the single mainactivity mordern approach by combining the code in the MainActivity.kt. Added the Api service for the login and otp verification added the Network folder.
Added the Functionality TransferMoney.kt which has two types MyAccount and AnotherAccount.

Change the navigation way in all the files in which have the navigation.

Added The Conformation screen For SendMoneyConfirmationSheet.kt and TransferMyAccConformationSheet.kt and TransferToMemberConfirmationScreen.kt.

Added the Payment Summary screen for the TransferMyAccConfirmationSheet.kt added this TransferSuccessScreen.kt for summary with the recipt.